+++ src/main/resources/potatologistics.mixins.json
@@ -4,7 +4,7 @@
   "package": "deboni.potatologistics.mixin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "RenderBLocksMixin"
+    "RenderBlocksMixin",
   ],
   "client": [
   ],
